Move field struct to separate file
