+++ src/nsif/mod.rs
@@ -1,6 +1,7 @@
 use std::{fmt::Display, fs::File};
 
-mod fileheader;
+pub mod field;
+pub mod fileheader;
 
 use fileheader::FileHeader;
 
@@ -55,3 +56,18 @@ impl Display for NSIF {
         write!(f, "{}", self.file_header)
     }
 }
+
+pub fn parse_string(vec: &Vec<u8>) -> Result<String, Box<dyn std::error::Error>> {
+    String::from_utf8(vec.clone()).map_err(Into::into)
+}
+
+/*
+fn parse_number(vec: &Vec<u8>) -> Result<i32, Box<dyn std::error::Error>> {
+    let s = parse_string(vec)?.trim_start_matches('0').to_owned();
+    if s.is_empty() {
+        Ok(0)
+    } else {
+        s.parse::<i32>().map_err(Into::into)
+    }
+}
+*/
